chore: move node builder trait helpers to separate module
